permissions

+++ pages/impl/src/main/java/org/sakaiproject/pages/impl/PagesServiceImpl.java
@@ -1,7 +1,11 @@
 package org.sakaiproject.pages.impl;
 
+import org.sakaiproject.authz.api.FunctionManager;
+import org.sakaiproject.authz.api.SecurityService;
+import org.sakaiproject.pages.api.PagesPermissionException;
 import org.sakaiproject.pages.api.PagesService;
 import org.sakaiproject.pages.api.PageTransferBean;
+import org.sakaiproject.pages.api.Permissions;
 import org.sakaiproject.pages.api.repository.PageRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,9 +17,23 @@ import java.util.stream.Collectors;
 
 public class PagesServiceImpl implements PagesService {
 
+    @Autowired FunctionManager functionManager;
+
     @Autowired PageRepository pageRepository;
 
-    public PageTransferBean savePage(PageTransferBean bean) {
+    @Autowired SecurityService securityService;
+
+    public void init() {
+
+        functionManager.registerFunction(Permissions.ADD_PAGE, true);
+        functionManager.registerFunction(Permissions.DELETE_PAGE, true);
+    }
+
+    public PageTransferBean savePage(PageTransferBean bean) throws PagesPermissionException {
+
+        if (!securityService.unlock(Permissions.ADD_PAGE, "/site/" + bean.siteId)) {
+            throw new PagesPermissionException();
+        }
 
         System.out.println(bean.asPage().getId());
 
@@ -43,7 +61,11 @@ public class PagesServiceImpl implements PagesService {
         return pageRepository.findById(pageId).map(PageTransferBean::of);
     }
 
-    public void deletePage(String pageId, String siteId) {
+    public void deletePage(String pageId, String siteId) throws PagesPermissionException {
+
+        if (!securityService.unlock(Permissions.DELETE_PAGE, "/site/" + siteId)) {
+            throw new PagesPermissionException();
+        }
 
         pageRepository.deleteById(pageId);
     }
